Remove unused dep
