--- examples/blinky/.cargo/config.toml
@@ -1,11 +1,0 @@
-[target.'cfg(all(target_arch = "arm", target_os = "none"))']
-# replace STM32F429ZITx with your chip as listed in `probe-run --list-chips`
-#runner = "probe-run --chip STM32L475VGT6"
-runner = "probe-run --chip STM32L475VG"
-#runner = "probe-run --chip STM32L4S5VI"
-
-[build]
-target = "thumbv7em-none-eabi"
-
-[env]
-DEFMT_LOG = "trace"
