+++ examples/blinky/Cargo.toml
@@ -20,7 +20,6 @@ panic-probe = { version = "0.3", features = ["print-defmt"], optional = true }
 defmt-rtt = { version = "0.3", optional = true }
 defmt = "0.3"
 
-[build-dependencies]
 drogue-metaboard = { path = "../metaboard", default-features = false }
 
 [features]
@@ -28,6 +27,6 @@ drogue-metaboard = { path = "../metaboard", default-features = false }
 "board+nrf52-dk" = ["embassy-nrf/nrf52832", "panic-probe", "defmt-rtt", "drogue-metaboard/board+nrf52-dk"]
 
 [patch.crates-io]
-embassy-executor = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0", default-features = false, features = ["nightly", "integrated-timers"] }
-embassy-stm32 = { git = "https://github.com/embassy-rs/embassy", tag = "embassy-time-v0.1.0", optional = true, features = ["exti", "unstable-traits", "nightly", "time-driver-any"] }
-embassy-nrf = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0", optional = true, features = ["unstable-traits", "nightly"] }
+embassy-executor = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0" }
+embassy-stm32 = { git = "https://github.com/embassy-rs/embassy", tag = "embassy-time-v0.1.0" }
+embassy-nrf = { git = "https://github.com/embassy-rs/embassy.git", tag = "embassy-time-v0.1.0" }
