+++ examples/blinky/build.rs
@@ -1,14 +1,4 @@
-use std::{env, fs::File, io::Write, path::PathBuf};
-
 fn main() {
-    let out = &PathBuf::from(env::var_os("OUT_DIR").unwrap());
-    let mut v = Vec::new();
-    for (key, value) in env::vars() {
-        v.push(format!("{}={}", key, value));
-    }
-
-    println!("cargo:rustc-link-search={}", out.display());
-    //    println!("cargo:rerun-if-changed=memory.x");
     println!("cargo:rustc-link-arg-bins=--nmagic");
     println!("cargo:rustc-link-arg-bins=-Tlink.x");
     println!("cargo:rustc-link-arg-bins=-Tdefmt.x");
