--- examples/metaboard-gen/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "drogue-metaboard-gen"
-version = "0.1.0"
