+++ examples/metaboard/Cargo.toml
@@ -2,9 +2,6 @@
 name = "drogue-metaboard"
 version = "0.1.0"
 
-[build-dependencies]
-drogue-metaboard-gen = { path = "../metaboard-gen" }
-
 [features]
 "board+b-l475e-iot01a" = []
 "board+nrf52-dk" = []
