+++ examples/metaboard/src/lib.rs
@@ -1,0 +1,1 @@
+#![no_std]
